third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/background-position-expected.txt

This is a testharness.js-based test.
FAIL Can set 'background-position' to CSS-wide keywords assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to var() references assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
FAIL Can set 'background-position' to a position Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
PASS Setting 'background-position' to a length throws TypeError
PASS Setting 'background-position' to a percent throws TypeError
PASS Setting 'background-position' to a time throws TypeError
PASS Setting 'background-position' to an angle throws TypeError
PASS Setting 'background-position' to a flexible length throws TypeError
PASS Setting 'background-position' to a number throws TypeError
PASS Setting 'background-position' to a URL throws TypeError
PASS Setting 'background-position' to a transform throws TypeError
Harness: the test ran to completion.
